+++ zulu/parser.py
@@ -263,7 +263,8 @@ def format_timedelta(delta,
                      granularity='second',
                      threshold=0.85,
                      add_direction=False,
-                     format='long'):
+                     format='long',
+                     locale=LC_TIME):
     """Return timedelta as a formatted string.
 
     Args:
@@ -278,6 +279,8 @@ def format_timedelta(delta,
             `'in 1 hour'`). Defaults to ``False``.
         format (str, optional): Can be one of "long", "short", or "narrow".
             Defaults to `'long`'.
+        locale (str|Locale, optional): A ``Locale`` object or locale
+            identifer. Defaults to system default.
     """
     if granularity not in TIMEDELTA_GRANULARITIES:
         grans = ', '.join('"{0}"'.format(gra)
@@ -296,7 +299,8 @@ def format_timedelta(delta,
                              granularity=granularity,
                              threshold=threshold,
                              add_direction=add_direction,
-                             format=format)
+                             format=format,
+                             locale=locale)
 
 
 def get_timezone(tz):