+++ zulu/timedelta.py
@@ -6,6 +6,8 @@ from __future__ import absolute_import
 
 from datetime import timedelta
 
+from babel.dates import LC_TIME
+
 from . import parser
 
 
@@ -39,7 +41,8 @@ class Delta(timedelta):
                granularity='second',
                threshold=0.85,
                add_direction=False,
-               format='long'):
+               format='long',
+               locale=LC_TIME):
         """Return timedelta as a formatted string.
 
         Args:
@@ -54,12 +57,15 @@ class Delta(timedelta):
                 `'in 1 hour'`). Defaults to ``False``.
             format (str, optional): Can be one of "long", "short", or "narrow".
                 Defaults to `'long`'.
+            locale (str|Locale, optional): A ``Locale`` object or locale
+                identifer. Defaults to system default.
         """
         return parser.format_timedelta(self,
                                        granularity=granularity,
                                        threshold=threshold,
                                        add_direction=add_direction,
-                                       format=format)
+                                       format=format,
+                                       locale=locale)
 
     def __repr__(self):  # pragma: no cover
         """Return representation of :class:`.Delta`."""